refactor(context): extract session module
